put loader in edit page and action in edit-form component
