+++ event-calendar/app/routes/admin.categories.new.tsx
@@ -15,30 +15,38 @@ export const validator = withZod(
 );
 
 export const action = async ({ request }: DataFunctionArgs) => {
-  const data = await validator.validate(await request.formData());
-  if (data.error) return validationError(data.error);
-  const { name, color } = data.data;
+  const formData = await request.formData();
+  const action = formData.get("action");
 
-  try {
-    await createCategory({
-      name,
-      color,
-    });
+  if (action === "create") {
+    // Your code for handling the "create" action goes here
+    const data = await validator.validate(formData);
+    if (data.error) return validationError(data.error);
+    const { name, color } = data.data;
 
-    return redirect("/admin/categories");
-  } catch (error) {
-    console.error("Error creating category", error);
+    try {
+      await createCategory({
+        name,
+        color,
+      });
 
-    return json(
-      {
-        title: "Error",
-        details: "An error occurred while creating the category.",
-      },
-      { status: 500 }
-    );
-  } finally {
-    await db.$disconnect();
+      return redirect("/admin/categories");
+    } catch (error) {
+      console.error("Error creating category", error);
+
+      return json(
+        {
+          title: "Error",
+          details: "An error occurred while creating the category.",
+        },
+        { status: 500 }
+      );
+    } finally {
+      await db.$disconnect();
+    }
   }
+
+  // Your code for handling other actions goes here
 };
 
 function addCategory() {
